Update Docker support.

+++ docker/build.py
@@ -2,19 +2,24 @@
 
 import os
 import sys
+import facts
 import docker
 
 def main():
+    if not docker.DockerService.is_running():
+        docker.DockerService.start()
+
     try:
         docker.DockerfileBuilder() \
             .with_template('./Dockerfile.tpl') \
-            .with_vars(docker.get_user_variables()) \
+            .with_vars(facts.get_facts()['user']) \
             .with_output('./Dockerfile') \
             .build()
 
         docker.Docker().build(path='./', tag='arwmar/images:base')
     finally:
-        os.unlink('Dockerfile')
+        if os.path.exists('Dockerfile'):
+           os.unlink('Dockerfile')
 
 if __name__ == '__main__':
     sys.exit(main())
